separated pyvista and matplotlib helpers into different files
