+++ py-engeom/examples/align_points_to_mesh_constrained.py
@@ -5,7 +5,7 @@ from pyvista import Plotter
 
 from engeom.align import points_to_mesh, Dof6
 from engeom.geom3 import Mesh3, Iso3
-from engeom.plot import PyvistaPlotterHelper
+from engeom.plot.pyvista import PyvistaPlotterHelper
 
 
 def main():
