+++ py-engeom/examples/display_mesh.py
@@ -1,6 +1,6 @@
 from pyvista import Plotter
 from engeom.geom3 import Mesh3
-from engeom.plot import PyvistaPlotterHelper
+from engeom.plot.pyvista import PyvistaPlotterHelper
 
 
 def main():
