--- py-engeom/python/engeom/_plot/__init__.py
@@ -1,1 +1,0 @@
-from .common import LabelPlace
